Made person appear infront of hazmat suit, added moveSpeed, and 2d array of booleans named walls
moveSpeed determins how many pixels person will move each time a key is pressed. Put it in one place so that it can easily be changed if needed to.
walls, the 2D array of booleans, was added to determine where the walls where be in the stage. Still needs to be implemented. Is currently just instantiated.